refactor: Move `getTranslationsFromMultisigRequest` to own file
